pyproject.toml
[project]
name = "cloudredux"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.23.0",
    "google-cloud-aiplatform>=1.136.0",
    "litellm>=1.81.8",
    "pydantic>=2.12.5",
    "ruff>=0.15.0",
    "streamlit>=1.54.0",
]
